Extract matchers with side effects support
