+++ spec/rubocop/cop/rspec/aggregate_examples_spec.rb
@@ -1,8 +1,6 @@
 RSpec.describe RuboCop::Cop::RSpec::AggregateExamples, :config do
   subject(:cop) { described_class.new(config) }
 
-  let(:cop_config) { {} }
-
   shared_examples 'detects and autocorrects in example group' do |group|
     context "with '#{group}'" do
       offensive_source = <<-RUBY
@@ -333,78 +331,6 @@ RSpec.describe RuboCop::Cop::RSpec::AggregateExamples, :config do
     it 'detects and autocorrects' do
       expect_offense(offensive_source)
       expect_correction(good_source)
-    end
-  end
-
-  context 'with matchers with side effects' do
-    context 'without no side effect matchers defined in configuration' do
-      let(:cop_config) do
-        { 'MatchersWithSideEffects' => [] }
-      end
-
-      offensive_source = <<-RUBY
-        describe do
-          it { expect(entry).to validate_absence_of(:comment) }
-          it { expect(entry).to validate_presence_of(:description) }
-          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Aggregate with the example at line 2.
-        end
-      RUBY
-
-      good_source = <<-RUBY
-        describe do
-          specify do
-            expect(entry).to validate_absence_of(:comment)
-            expect(entry).to validate_presence_of(:description)
-          end
-        end
-      RUBY
-
-      it 'detects and autocorrects' do
-        expect_offense(offensive_source)
-        expect_correction(good_source)
-      end
-    end
-
-    context 'with the default configuration' do
-      offensive_source = <<-RUBY
-        describe do
-          it { expect(entry).to validate_absence_of(:comment) }
-          it { expect(entry).to validate_presence_of(:description) }
-          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Aggregate with the example at line 2. IMPORTANT! Pay attention to the expectation order, some of the matchers have side effects.
-        end
-      RUBY
-
-      it 'detects an offense, but does not correct it' do
-        expect_offense(offensive_source)
-        expect_no_corrections
-      end
-    end
-
-    context 'with mixed matchers' do
-      offensive_source = <<-RUBY
-        describe 'with and without side effects' do
-          it { expect(fruit).to be_good }
-          it { expect(fruit).to be_cheap }
-          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Aggregate with the example at line 2.
-          it { expect(fruit).to validate_presence_of(:color) }
-          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Aggregate with the example at line 2. IMPORTANT! Pay attention to the expectation order, some of the matchers have side effects.
-        end
-      RUBY
-
-      good_source = <<-RUBY
-        describe 'with and without side effects' do
-          specify do
-            expect(fruit).to be_good
-            expect(fruit).to be_cheap
-          end
-          it { expect(fruit).to validate_presence_of(:color) }
-        end
-      RUBY
-
-      it 'detects an offense in offensive_source code' do
-        expect_offense(offensive_source)
-        expect_correction(good_source)
-      end
     end
   end
